### 2025-03-21 #7935 @editorialbot
[PRE REVIEW]: Turftopic: Topic Modelling with Contextual Representations from Sentence Transformers
**Submitting author:** @x-tabdeveloping (<a href="http://orcid.org/0000-0001-9652-4498">Márton Kardos</a>)
**Repository:** https://github.com/x-tabdeveloping/turftopic
**Branch with paper.md** (empty if default branch):
**Version:** v0.14.1
