### 2025-03-30 14d6d45
Added meaningful errors to KeyNMF
