+++ turftopic/models/_keynmf.py
@@ -8,8 +8,12 @@ import igraph as ig
 import numpy as np
 import scipy.sparse as spr
 from sklearn.base import clone
-from sklearn.decomposition._nmf import (NMF, MiniBatchNMF, _initialize_nmf,
-                                        _update_coordinate_descent)
+from sklearn.decomposition._nmf import (
+    NMF,
+    MiniBatchNMF,
+    _initialize_nmf,
+    _update_coordinate_descent,
+)
 from sklearn.exceptions import NotFittedError
 from sklearn.feature_extraction.text import CountVectorizer
 from sklearn.metrics.pairwise import cosine_similarity
@@ -191,7 +195,7 @@ class SBertKeywordExtractor:
             self._add_terms(new_terms)
         total = embeddings.shape[0]
         # Relevance based on similarity to seed embedding
-        document_relevance = None
+        document_relevance = np.ones(embeddings.shape[0])
         if seed_embedding is not None:
             if self.metric == "cosine":
                 document_relevance = cosine_similarity(
@@ -225,8 +229,7 @@ class SBertKeywordExtractor:
             else:
                 sim = np.dot(word_embeddings, embedding[0]).T
             # If a seed is specified, we multiply by the document's relevance
-            if document_relevance is not None:
-                sim = document_relevance[i] * sim
+            sim = document_relevance[i] * sim
             kth = min(self.top_n, len(sim) - 1)
             top = np.argpartition(-sim, kth)[:kth]
             top_words = batch_vocab[important_terms][top]
