### 2025-03-30 86bc02d
Added base model for multimodal topic models
