### 2025-03-31 f1cea06
Added plotting and elementary multimodal modeling to S3
