+++ turftopic/multimodal.py
@@ -1,16 +1,14 @@
 from __future__ import annotations
 
 from abc import abstractmethod
-from typing import Iterable
+from typing import Iterable, Optional, TypedDict
 
 import numpy as np
 from PIL import Image
 
-from turftopic.base import ContextualModel
-
 UrlStr = str
 
-ImageRepr = [Image | UrlStr]
+ImageRepr = [Image.Image | UrlStr]
 
 
 def _load_images(images: Iterable[ImageRepr]) -> Iterable[Image]:
@@ -27,7 +25,13 @@ def _naive_join_embeddings(
     return np.mean(np.stack([text_embeddings, image_embeddings]), axis=0)
 
 
-class MultimodalModel(ContextualModel):
+class MultimodalEmbeddings(TypedDict):
+    text_embeddings: np.ndarray
+    image_embeddings: np.ndarray
+    document_embeddings: np.ndarray
+
+
+class MultimodalModel:
     """Base model for multimodal topic models."""
 
     def encode_multimodal(
@@ -49,7 +53,9 @@ class MultimodalModel(ContextualModel):
                 self.encoder_.get_image_embeddings(images)
             )
         else:
-            image_embeddings = self.encoder_.encode(images)
+            image_embeddings = np.stack(
+                [self.encoder_.encode(image) for image in images]
+            )
         if hasattr(self.encoder_, "get_fused_embeddings"):
             document_embeddings = np.array(
                 self.encoder_.get_fused_embeddings(
@@ -71,8 +77,95 @@ class MultimodalModel(ContextualModel):
     @abstractmethod
     def fit_transform_multimodal(
         self,
-        raw_documents: Iterable[str],
-        images: Iterable[tuple[ImageRepr] | ImageRepr],
+        raw_documents: list[str],
+        images: list[ImageRepr],
         y=None,
+        embeddings: Optional[MultimodalEmbeddings] = None,
     ) -> np.ndarray:
         pass
+
+    def _top_image_grid(self, ind_topic: int, final_size=(1200, 1200)):
+        grid_size = (3, 3)
+        grid_img = Image.new("RGB", final_size, (255, 255, 255))
+        cell_width = final_size[0] // grid_size[0]
+        cell_height = final_size[1] // grid_size[1]
+        for idx, img in enumerate(self.top_images[ind_topic]):
+            img = img.resize(
+                (cell_width, cell_height), resample=Image.Resampling.LANCZOS
+            )
+            x_offset = (idx % grid_size[0]) * cell_width
+            y_offset = (idx // grid_size[1]) * cell_height
+            grid_img.paste(img, (x_offset, y_offset))
+        return grid_img
+
+    def plot_topics_with_images(self, n_columns: int = 3):
+        try:
+            import plotly.graph_objects as go
+        except (ImportError, ModuleNotFoundError) as e:
+            raise ModuleNotFoundError(
+                "Please install plotly if you intend to use plots in Turftopic."
+            ) from e
+        fig = go.Figure()
+        width, height = 1200, 1200
+        scale_factor = 0.25
+        w, h = width * scale_factor, height * scale_factor
+        padding = 10
+        n_cols = 3
+        n_components = self.components_.shape[0]
+        n_rows = n_components // n_cols + int(bool(n_components % n_cols))
+        figure_height = (h + padding) * n_rows
+        figure_width = (w + padding) * n_cols
+        fig = fig.add_trace(
+            go.Scatter(
+                x=[0, figure_width],
+                y=[0, figure_height],
+                mode="markers",
+                marker_opacity=0,
+            )
+        )
+        vocab = self.get_vocab()
+        for i, component in enumerate(self.components_):
+            col = i % n_cols
[…]
