### 2025-03-31 bedd226
Refined multimodal encoder interface
