### 2025-03-31 9d8e30e
Refined multimodal encoder interface
