### 2025-03-31 26a313c
Added multimodal encoder interface
