+++ turftopic/models/keynmf.py
@@ -274,11 +274,9 @@ class KeyNMF(ContextualModel, DynamicTopicModel, MultimodalModel):
             console.log("Model fitting done.")
             status.update("Transforming images")
             self.image_topic_matrix = self.model.transform(image_keywords)
-            self.top_images: list[list[Image.Image]] = []
-            for image_topic_vector in self.image_topic_matrix.T:
-                top_im_ind = np.argsort(-image_topic_vector)[:9]
-                top_im = [images[i] for i in top_im_ind]
-                self.top_images.append(top_im)
+            self.top_images: list[list[Image.Image]] = self.collect_top_images(
+                images, self.image_topic_matrix
+            )
             console.log("Images transformed")
         self.document_topic_matrix = doc_topic_matrix
         self.document_term_matrix = self.model.vectorize(document_keywords)
