### 2025-04-01 8284aaf
Added negative images to s3
