### 2025-04-01 cb793b6
Added negative image plot to S3
