+++ turftopic/models/decomp.py
@@ -228,6 +228,109 @@ class SemanticSignalSeparation(
             console.log("Images transformed")
         return doc_topic
 
+    def plot_topics_with_images(self, n_columns: int = 3, grid_size: int = 4):
+        try:
+            import plotly.graph_objects as go
+        except (ImportError, ModuleNotFoundError) as e:
+            raise ModuleNotFoundError(
+                "Please install plotly if you intend to use plots in Turftopic."
+            ) from e
+        fig = go.Figure()
+        width, height = 1200, 1200
+        scale_factor = 0.25
+        w, h = width * scale_factor, height * scale_factor
+        padding = 10
+        figure_height = (h + padding) * self.n_components
+        figure_width = (w + padding) * 2
+        fig = fig.add_trace(
+            go.Scatter(
+                x=[0, figure_width],
+                y=[0, figure_height],
+                mode="markers",
+                marker_opacity=0,
+            )
+        )
+        vocab = self.get_vocab()
+        for i, component in enumerate(self.components_):
+            positive = vocab[np.argsort(-component)[:7]]
+            negative = vocab[np.argsort(component)[:7]]
+            pos_image = self._image_grid(
+                self.top_images[i],
+                (width, height),
+                grid_size=(grid_size, grid_size),
+            )
+            neg_image = self._image_grid(
+                self.negative_images[i],
+                (width, height),
+                grid_size=(grid_size, grid_size),
+            )
+            x0 = 0
+            y0 = (h + padding) * (self.n_components - i)
+            fig = fig.add_layout_image(
+                dict(
+                    x=x0,
+                    sizex=w,
+                    y=y0,
+                    sizey=h,
+                    xref="x",
+                    yref="y",
+                    opacity=1.0,
+                    layer="below",
+                    sizing="stretch",
+                    source=pos_image,
+                ),
+            )
+            fig.add_annotation(
+                x=(w / 2),
+                y=(h + padding) * (self.n_components - i) - (h / 2),
+                text="<b> " + "<br> ".join(positive),
+                font=dict(
+                    size=16,
+                    family="Times New Roman",
+                    color="white",
+                ),
+                bgcolor="rgba(0,0,255, 0.5)",
+            )
+            x0 = (w + padding) * 1
+            fig = fig.add_layout_image(
+                dict(
+                    x=x0,
+                    sizex=w,
+                    y=y0,
+                    sizey=h,
+                    xref="x",
+                    yref="y",
+                    opacity=1.0,
+                    layer="below",
+                    sizing="stretch",
+                    source=neg_image,
+                ),
+            )
+            fig.add_annotation(
+                x=(w + padding) + (w / 2),
+                y=(h + padding) * (self.n_components - i) - (h / 2),
+                text="<b> " + "<br> ".join(negative),
+                font=dict(
+                    size=16,
+                    family="Times New Roman",
+                    color="white",
+                ),
+                bgcolor="rgba(255,0,0, 0.5)",
+            )
+        fig = fig.update_xaxes(visible=False, range=[0, figure_width])
+        fig = fig.update_yaxes(
+            visible=False,
+            range=[0, figure_height],
+            # the scaleanchor attribute ensures that the aspect ratio stays constant
+            scaleanchor="x",
+        )
+        fig = fig.update_layout(
+            width=figure_width,
+            height=figure_height,
+            margin={"l": 0, "r": 0, "t": 0, "b": 0},
+        )
+        return fig
+
     def _rename_automatic(self, namer: TopicNamer) -> list[str]:
         """Names topics with a topic namer in the model.
 
