### 2025-04-01 3269ed3
Added multimodal embedding structure validation
