+++ turftopic/models/ctm.py
@@ -290,6 +290,7 @@ class AutoEncodingTopicModel(ContextualModel, MultimodalModel):
         y=None,
         embeddings: Optional[MultimodalEmbeddings] = None,
     ) -> np.ndarray:
+        self.validate_embeddings(embeddings)
         console = Console()
         self.multimodal_embeddings = embeddings
         with console.status("Fitting model") as status:
