+++ turftopic/models/gmm.py
@@ -142,6 +142,7 @@ class GMM(ContextualModel, DynamicTopicModel, MultimodalModel):
         y=None,
         embeddings: Optional[MultimodalEmbeddings] = None,
     ) -> np.ndarray:
+        self.validate_embeddings(embeddings)
         console = Console()
         self.multimodal_embeddings = embeddings
         with console.status("Fitting model") as status:
