+++ turftopic/models/decomp.py
@@ -176,6 +176,7 @@ class SemanticSignalSeparation(
         y=None,
         embeddings: Optional[MultimodalEmbeddings] = None,
     ) -> np.ndarray:
+        self.validate_embeddings(embeddings)
         console = Console()
         self.multimodal_embeddings = embeddings
         with console.status("Fitting model") as status:
