### 2025-04-01 fe2c87b
Fixed multimodal interface types so that they work with 3.9
