+++ turftopic/encoders/multimodal.py
@@ -1,8 +1,6 @@
 from typing import Protocol
 
 from PIL import Image
-from torch import Tensor
-from torch.utils.data import DataLoader
 
 
 class MultimodalEncoder(Protocol):
@@ -14,15 +12,15 @@ class MultimodalEncoder(Protocol):
         *,
         batch_size: int = 8,
         **kwargs,
-    ) -> Tensor: ...
+    ): ...
 
     def get_image_embeddings(
         self,
-        images: list[Image.Image] | DataLoader,
+        images: list[Image.Image],
         *,
         batch_size: int = 8,
         **kwargs,
-    ) -> Tensor: ...
+    ): ...
 
     def get_fused_embeddings(
         self,
@@ -30,4 +28,4 @@ class MultimodalEncoder(Protocol):
         images: list[Image.Image] = None,
         batch_size: int = 8,
         **kwargs,
-    ) -> Tensor: ...
+    ): ...
