### 2025-04-02 4a88a00
Added validation to encoder model in multimodal models
