+++ turftopic/models/cluster.py
@@ -208,6 +208,7 @@ class ClusteringTopicModel(
             self.encoder_ = SentenceTransformer(encoder)
         else:
             self.encoder_ = encoder
+        self.validate_encoder()
         if vectorizer is None:
             self.vectorizer = default_vectorizer()
         else:
