+++ turftopic/models/ctm.py
@@ -159,6 +159,7 @@ class AutoEncodingTopicModel(ContextualModel, MultimodalModel):
             self.encoder_ = SentenceTransformer(encoder)
         else:
             self.encoder_ = encoder
+        self.validate_encoder()
         if vectorizer is None:
             self.vectorizer = default_vectorizer()
         else:
