+++ turftopic/models/gmm.py
@@ -87,6 +87,7 @@ class GMM(ContextualModel, DynamicTopicModel, MultimodalModel):
             self.encoder_ = SentenceTransformer(encoder)
         else:
             self.encoder_ = encoder
+        self.validate_encoder()
         if vectorizer is None:
             self.vectorizer = default_vectorizer()
         else:
