### 2025-04-03 9f990ae
Multimodal models now encode None images as np.nan vectors
