### 2025-04-03 81ed6a7
Added multimodal test
