### 2025-04-03 4166909
Version bump, added pillow as a dependency
