+++ pyproject.toml
@@ -9,7 +9,7 @@ profile = "black"
 
 [tool.poetry]
 name = "turftopic"
-version = "0.14.1"
+version = "0.15.0"
 description = "Topic modeling with contextual representations from sentence transformers."
 authors = ["Márton Kardos <power.up1163@gmail.com>"]
 license = "MIT"
@@ -26,6 +26,7 @@ rich = "^13.6.0"
 huggingface-hub = ">=0.23.2"
 joblib = "^1.2.0"
 igraph = "~0.11.6"
+pillow = "~10.4.0"
 snowballstemmer = {version=">=2.0.0", optional=true}
 spacy = {version=">=3.6.0", optional=true}
 jieba = {version=">=0.40.0", optional=true}
