### 2025-04-09 3d2d7d1
Added docstrings for multimodal base model
