### 2025-04-09 f978a01
Rephrase BETA note
