### 2025-04-19 #7935 @abhishektiwari
@x-tabdeveloping I will be helping to edit this article. Before we can move, can you please action following?

- [ ]  Add Community/contribution guidelines to your project

- [ ]  Remove unused citation in `paper.bib`

```
blei_dynamic
blei_hierarchical
blei_prob_topic_models
ctm_docs
prodla
```

- [ ]  Please add `References` h1 at end of `paper.md`

```
# References
```

- [ ]  Double check DOI for some of the references.

E.g. _BERTopic: Neural topic modeling with a class-based TF-IDF procedure_ DOI
[10.48550/arXiv.2203.05794](https://doi.org/10.48550/arXiv.2203.05794), _Top2Vec: Distributed Representations of Topics_  DOI [10.48550/arXiv.2008.09470](https://doi.org/10.48550/arXiv.2008.09470)

Once updated me please let me know.
