### 2025-04-19 #7935 @abhishektiwari
@x-tabdeveloping You could create a CONTRIBUTING.md file (see here for some examples: https://contributing.md/example/) and link it back to your README.
