### 2025-04-20 08fc957
Added contribution guidelines and code of conduct
