### 2025-04-24 #7935 @abhishektiwari
Thanks @x-tabdeveloping I will suggest to add arXiv DOI where applicable. E.g. Top2Vec and BERTopic.

Also, can you suggest a few potential reviewers? You can also share GitHub handles of potential reviewers. Please be sure not to @-tag your suggestions, to avoid unwanted GitHub notifications to potential reviewers.
