### 2025-04-25 bdf4fba
Added arxiv doi where applicable
