+++ paper.bib
@@ -6,6 +6,7 @@
       archivePrefix={arXiv},
       primaryClass={cs.LG},
       url={https://arxiv.org/abs/2406.09556}, 
+      doi = "10.48550/arXiv.2406.09556",
 }
 
 @inproceedings{keynmf,
@@ -20,6 +21,7 @@
     series = "CEUR Workshop Proceedings",
     publisher = "CEUR-WS",
     pages = "829--847",
+    doi="10.48550/arXiv.2410.12791",
     editor = "Haverals, {Wouter } and Koolen, {Marijn } and Thompson, {Laure }",
     booktitle = "Proceedings of the Computational Humanities Research Conference 2024",
     address = "Germany",
@@ -29,6 +31,7 @@
   title={BERTopic: Neural topic modeling with a class-based TF-IDF procedure},
   author={Grootendorst, Maarten},
   journal={arXiv preprint arXiv:2203.05794},
+  doi="10.48550/arXiv.2203.05794",
   year={2022}
 }
 
@@ -123,7 +126,8 @@
       year={2020},
       eprint={2008.09470},
       archivePrefix={arXiv},
-      primaryClass={cs.CL}
+      primaryClass={cs.CL},
+      doi="10.48550/arXiv.2008.09470",
 }
 
 @article{scikit-learn,
@@ -143,6 +147,7 @@
     title={FASTopic: Pretrained Transformer is a Fast, Adaptive, Stable, and Transferable Topic Model},
     author={Wu, Xiaobao and Nguyen, Thong Thanh and Zhang, Delvin Ce and Wang, William Yang and Luu, Anh Tuan},
     booktitle={The Thirty-eighth Annual Conference on Neural Information Processing Systems},
+    doi="10.48550/arXiv.2405.17978",
     year={2024}
 }
 
