### 2025-04-25 #7935 @abhishektiwari
👋 Hi @PetrKorab @ArneTillmann @jasonnance, would any of you be willing to review this submission for JOSS?  The submission under consideration is "Turftopic: Topic Modelling with Contextual Representations from Sentence Transformers".

Based on your past authorship/reviews of related JOSS publications, I think you might be able to provide a great review of this submission. If not, could you recommend any potential reviewers?

Thanks
Abhi
