### 2025-05-03 #7935 @abhishektiwari
👋 Hi @jasonnance @jamiehadd @marccanby @cerule7 would any of you be willing to review this submission for JOSS? The submission under consideration is "Turftopic: Topic Modelling with Contextual Representations from Sentence Transformers".

We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:

https://joss.readthedocs.io/en/latest/review_criteria.html.

Based on your experience, we think you might be able to provide a great review of this submission.

If not, could you recommend any potential reviewers?

Thanks Abhi
