### 2025-05-05 3b6a059
Added convenience methods for collecting top documents, words and images to TopicData
