### 2025-05-05 9f9e92d
Made get_top_images conform to the same interface
