### 2025-05-05 9e0dec6
Moved image plotting functionality to TopicContainer
