### 2025-05-05 fe56cd3
Added multimodal TopicData
