### 2025-05-05 cb99e9b
Added n_components to container
