### 2025-05-05 ac047fa
Fixed issues with self.n_components
