### 2025-05-06 2d555bd
Fixed bug with multimodal encoders
