### 2025-05-09 #7935 @marccanby
I am also happy to review! Please let me know what I need to do.

--
Marc Canby

On May 9, 2025 at 1:45:04 PM, Abhishek Tiwari ***@***.***)
wrote:

> *abhishektiwari* left a comment (openjournals/joss-reviews#7935)
> <https://github.com/openjournals/joss-reviews/issues/7935#issuecomment-2867448725>
>
> Thanks @Jemoka <https://github.com/Jemoka> for agreeing to review it. I
> will wait for a day or two for others to respond before we start review at
> new Github issue.
>
