### 2025-05-09 #8183 @editorialbot
[REVIEW]: Turftopic: Topic Modelling with Contextual Representations from Sentence Transformers
**Submitting author:** @x-tabdeveloping (<a href="http://orcid.org/0000-0001-9652-4498">Márton Kardos</a>)
**Repository:** https://github.com/x-tabdeveloping/turftopic
**Branch with paper.md** (empty if default branch):
**Version:** v.0.17.2
