### 2025-05-16 76bb035
Added multimodal topic browser
