+++ turftopic/container.py
@@ -1231,3 +1231,192 @@ class TopicContainer(ABC):
                 margin={"l": 0, "r": 0, "t": 0, "b": 0},
             )
             return fig
+
+    def plot_multimodal_topics(
+        self,
+        top_k: int = 10,
+        grid_size: int = 4,
+        raw_documents=None,
+        document_topic_matrix=None,
+    ):
+        """Plots all multimodal topics in a model along with top documents individually,
+        and provides a slider to switch between them.
+
+        Parameters
+        ----------
+        top_k: int = 10
+            Number of top words and documents to display.
+        grid_size: int, default 4
+            The square root of the number of images you want to display for a given topic.
+            For instance if grid_size==4, all topics will have 16 images displayed,
+            since the joint image will have 4 columns and 4 rows.
+        raw_documents: list of str, optional
+            List of documents to consider.
+        document_topic_matrix: ndarray of shape (n_documents, n_topics), optional
+            Document topic matrix to use. This is useful for transductive methods,
+            as they cannot infer topics from text.
+
+        """
+        if not hasattr(self, "top_images"):
+            raise ValueError(
+                "Model either has not been fit or was fit without images. top_images property missing."
+            )
+        try:
+            import plotly.express as px
+            import plotly.graph_objects as go
+            from plotly.subplots import make_subplots
+        except (ImportError, ModuleNotFoundError) as e:
+            raise ModuleNotFoundError(
+                "Please install plotly if you intend to use plots in Turftopic."
+            ) from e
+        negative_images = getattr(self, "negative_images", None)
+        negative_topics = (
+            self.get_top_words(top_k=top_k, positive=False)
+            if negative_images is not None
+            else None
+        )
+        specs = [{"type": "image"}, {"type": "table"}]
+        if negative_images is not None:
+            specs.append({"type": "image"})
+        fig = make_subplots(
+            rows=1,
+            cols=2 if negative_images is None else 3,
+            specs=[specs],
+            shared_yaxes=True,
+            shared_xaxes=True,
+        )
+        width, height = 1200, 1200
+        grid_size = 4
+        topics = self.get_top_words(top_k=top_k)
+        n_topics = len(topics)
+        annotations = []
+        for i, topic in enumerate(topics):
+            images = self.top_images[i]
+            image = TopicContainer._image_grid(
+                images, (width, height), grid_size=(grid_size, grid_size)
+            )
+            trace = px.imshow(image).data[0]
+            trace.visible = False
+            fig.add_trace(trace, col=1, row=1)
+            annt = dict(
+                x=width / 2,
+                y=height / 2,
+                text="<b> " + "<br> ".join(topic),
+                font=dict(
+                    size=16,
+                    family="Roboto Mono",
+                    color="white",
+                ),
+                bgcolor="rgba(0,0,255, 0.5)",
+                xref="x",
+                yref="y",
+            )
+            annotations.append(annt)
+        if negative_topics is not None:
+            for i, negative_topic in enumerate(negative_topics):
+                images = negative_images[i]
+                image = TopicContainer._image_grid(
+                    images, (width, height), grid_size=(grid_size, grid_size)
+                )
+                trace = px.imshow(image).data[0]
+                trace.visible = False
+                fig.add_trace(trace, col=3, row=1)
+                annotations.append(
+                    dict(
+                        x=width / 2,
+                        y=height / 2,
+                        text="<b> " + "<br> ".join(negative_topic),
+                        font=dict(
+                            size=16,
+                            family="Roboto Mono",
+                            color="white",
+                        ),
+                        bgcolor="rgba(255,0,0, 0.5)",
+                        xref="x2",
+                        yref="y2",
+                    )
+                )
[…]
