### 2025-05-16 98e2775
Switched to Roboto Mono on Hierarchical plots
