### 2025-05-16 324622f
Added image compass, removed size limitation on concept compass
