### 2025-05-19 684c0ed
Added convenience functions for Top2Vec and BERTopic
