+++ turftopic/__init__.py
@@ -1,6 +1,6 @@
 from turftopic.base import ContextualModel
 from turftopic.error import NotInstalled
-from turftopic.models.cluster import ClusteringTopicModel
+from turftopic.models.cluster import BERTopic, ClusteringTopicModel, Top2Vec
 from turftopic.models.decomp import SemanticSignalSeparation
 from turftopic.models.fastopic import FASTopic
 from turftopic.models.gmm import GMM
@@ -21,5 +21,7 @@ __all__ = [
     "AutoEncodingTopicModel",
     "ContextualModel",
     "FASTopic",
+    "Top2Vec",
+    "BERTopic",
     "load_model",
 ]
