### 2025-05-19 a204b4c
added UMAP as an optional dependency
