### 2025-05-19 f401750
Fixed inconsistent behaviour of label_binarize
