+++ docs/s3.md
@@ -176,13 +176,30 @@ pip install plotly
 You can display a compass based on a fitted model like so:
 
 ```python
-fig = model.concept_compass(topic_x=1, topic_y=4)
+fig = model.plot_concept_compass(topic_x=1, topic_y=4)
 fig.show()
 ```
 
 <figure>
   <img src="../images/arxiv_ml_compass.png" width="60%" style="margin-left: auto;margin-right: auto;">
   <figcaption> Concept Compass of ArXiv ML Papers along two semantic axes. </figcaption>
+</figure>
+
+#### Image Compass
+
+In multimodal contexts, you can also plot images along two chosen axes by using `plot_image_compass()`.
+
+```python
+model = SemanticSignalSeparation(10)
+model.fit_multimodal(corpus, images=images)
+
+fig = model.plot_image_compass(topic_x=0, topic_y=1)
+fig.show()
+```
+
+<figure>
+  <iframe src="../images/image_compass.html", title="Image Compass of IKEA furnitures along two semantic axes", style="height:800px;width:1200px;padding:0px;border:none;"></iframe>
+  <figcaption> Image Compass of IKEA furnitures along two semantic axes </figcaption>
 </figure>
 
 
