+++ mkdocs.yml
@@ -20,8 +20,8 @@ nav:
     - Semantic Signal Separation (S³): s3.md
     - KeyNMF: KeyNMF.md
     - GMM: GMM.md
-    - Clustering Models: clustering.md
-    - Autoencoding Models: ctm.md
+    - Clustering Models (BERTopic & Top2Vec): clustering.md
+    - Autoencoding Models (ZeroShotTM & CombinedTM): ctm.md
     - FASTopic: FASTopic.md
   - Encoders: encoders.md
   - Vectorizers: vectorizers.md
