### 2025-05-19 91d9f0c
Fixed mkdocs dependencies
