+++ pyproject.toml
@@ -34,9 +34,9 @@ openTSNE = {version="^1.0.0", optional=true}
 datamapplot = {version="^0.4.2", optional=true}
 pyro-ppl = { version = "^1.8.0", optional = true }
 openai = { version = "^1.40.0", optional = true }
-mkdocs = { version = "^1.5.2", optional = true }
-mkdocs-material = { version = "^9.5.12", optional = true }
-mkdocstrings = { version = "^0.24.0", extras = ["python"], optional = true }
+mkdocs = { version = "~1.5.3", optional = true }
+mkdocs-material = { version = "~9.5.12", optional = true }
+mkdocstrings = { version = "~0.24.0", extras = ["python"], optional = true }
 topic-wizard = { version = ">=1.0.0", optional = true}
 pytest = { version = ">=8.1.1", optional = true }
 umap-learn = { version = "^0.5.5", optional = true }
