### 2025-05-19 13d7ba7
fix typo in turftopic/utils.py
Co-authored-by: Kenneth Enevoldsen <kennethcenevoldsen@gmail.com>
