### 2025-05-19 #8183 @PetrKorab
@x-tabdeveloping JOSS is a journal for software with research applications. Please, 1) extend the **Use Cases** and **Target Audience** sections of the paper.  Discuss the applications in various disciplines, including economics and political science. Focus also on datasets where Turftopic can serve the researchers, and mention how Turftopic helps discover new insights. 2) Expand the Examples folder in the repo with the commented codes, adding a reference to datasets with a convenient data license. Thanks!

https://github.com/x-tabdeveloping/turftopic/issues/92
