### 2025-05-20 x-tabdeveloping/turftopic#93 @PetrKorab
Extend documentation - data preprocessing, bigrams (JOSS submission)
Please extend the documentation on data preprocessing:

* Topic models are sensitive to **data preprocessing**. Please extend the documentation on the preprocessing for each model. Do they need full preprocessing prior to training, or does this cause a bias? Can they be trained on lemmatized texts?

* In social sciences, where topic models are often used, **bigram** topic models help make important decisions better than single-word clustering results. Please mention in the documentation how researchers and analysts can train Turftopic to construct topics from bigrams. Some workarounds here include joining tokens with underscores if the models do not provide it.
