### 2025-05-20 #8183 @PetrKorab
@x-tabdeveloping
In social sciences, where topic models are often used, bigram topic models help make important decisions better than single-word clustering results. Please, add a section to the documentation on how researchers and analysts can train Turftopic to construct topics from bigrams. Some workarounds here include joining tokens with underscores if the models do not provide it.

Also, the documentation is really well-developed. Topic models are though sensitive to data preprocessing. Please also extend the documentation on the preprocessing for each model. Do they need full preprocessing prior to training, or does this cause a bias? Can they be trained on lemmatized texts?

These can be responded to in a separate section or for each model in the docs. Thanks!
https://github.com/x-tabdeveloping/turftopic/issues/93
