### 2025-05-20 x-tabdeveloping/turftopic#93 @x-tabdeveloping
Hi again @PetrKorab .

This is actually covered in the documentation, mainly under the [Vectorizers tab](https://x-tabdeveloping.github.io/turftopic/vectorizers/).
We discourage users from preprocessing data, since the sentence transformer's representations benefit from having access to the raw text.
We recommend recommend that people use different vectorizers to achieve such effects, which only filter, for which terms topic importance scores will be estimated, but do not actually remove anything from the text.
Contextual models are also much better at handling noise in data, so preprocessing is not as essential as with classical models most of the time.

Bigram topic modelling can for instance be easily achieved by passing a `CountVectorizer(ngram_range=(2,2))` to a Turftopic model.
We include lemmatizing, stemming and noun-phrase vectorizers as part of the package, as well as language-specific ones.

If you feel we should communicate this more clearly or in other parts of the documentation as well, your suggestions would be most welcome.
