+++ docs/tutorials/arxiv_ml.md
@@ -1,0 +1,245 @@
+# Building a Taxonomy of Machine Learning Papers
+
+Topic models can often be used to gain an overall understanding of what kinds of topics are being discussed in a corpus.
+If one wanted to build a taxonomy of the field of machine learning, a good starting point would be to investigate what sorts of machine learning articles have been put on ArXiv, the largest preprint server.
+
+In this tutorial, I will demonstrate how you can use Turftopic to discover themes in machine learning research using a clustering topic model.
+
+## Installation
+
+For this investigation we will use two additional libraries. `datasets` to fetch the dataset from HF Hub, as well as `plotly` for creating interactive plots of our results.
+We also need to install optional dependencies to be able to use umap-based clustering models, and then create a map using datamapplot.
+
+```bash
+pip install datasets plotly turftopic[umap-learn, datamapplot]
+```
+
+## Data Preparation
+
+We will download the dataset from HuggingFace Hub, using `datasets`.
+
+!!! note
+    In this example, we will use a subset of the data (10000 examples),
+    as it would otherwise take a long time to run on my laptop.
+    If you have a more powerful computer or a GPU at hand, I recommend running the topic model on the entire corpus.
+```python
+from datasets import load_dataset
+
+ds = load_dataset("CShorten/ML-ArXiv-Papers", split="train")
+# Subsampling dataset
+ds = ds.train_test_split(seed=42, test_size=10_000)["test"]
+abstracts = ds["abstract"]
+```
+
+Turftopic uses contextual embeddings of documents in order to understand what is in the corpus.
+We will use the small and fast `all-MiniLM-L6-v2` sentence transformer to produce embeddings of our data.
+
+!!! tip 
+    Sometimes you might want to reuse embeddings in different topic models, or save them to disk.
+    It is thus recommended, but not necessary to precompute them before running a topic model.
+
+```python
+from sentence_transformers import SentenceTransformer
+
+encoder = SentenceTransformer("all-MiniLM-L6-v2")
+embeddings = encoder.encode(abstracts, show_progress_bar=True)
+```
+
+## Model Training
+
+For building a hierarchy, we will assume that a paper is centered around one central theme, and we would like to categorize our articles into distinct groups.
+Clustering topic models, such as BERTopic and Top2Vec are well suited for this task, as they assume that a document belongs to a single cluster.
+They can also be used in settings, where we expect there to be a hierarchy of topics.
+This is certainly the case for machine learning, where, for instance, *auto-encoders* could be a sub-topic of *unsupervised learning*.
+
+In this example I am going to use the Top2Vec topic model, which discovers topic using UMAP and HDBSCAN, and produces very clean topic descriptions.
+Top2Vec learns the number of topics from the data, so we won't have to specify it a-priori.
+
+```python
+from turftopic import Top2Vec
+
+model = Top2Vec(encoder=encoder, random_state=42)
+topic_data = model.prepare_topic_data(abstracts, embeddings=embeddings)
+```
+
+## Interpreting Results
+
+Let's print the topic in our model in order to see what sorts of topic have been discovered.
+
+```python
+model.print_topics()
+```
+
+| Topic ID | Highest Ranking |
+| - | - |
+| -1 | softmax, supervised, cnn, cnns, autoencoders, learns, classifiers, imagenet, rnns, learning |
+| 0 | quantum, quantization, qubit, quantized, encoders, backpropagation, learns, disentangling, learning, decoders |
[…]
