+++ mkdocs.yml
@@ -15,6 +15,12 @@ nav:
     - Modifying and Finetuning Models: finetuning.md
     - Saving and Loading: persistence.md
     - Using TopicData: topic_data.md
+  - Tutorials:
+     - Tutorial Overview: tutorials/overview.md
+     - Analyzing the Landscape of Machine Learning Research: tutorials/arxiv_ml.md
+     - Discourse Analysis on Morality and Religion: tutorials/religious.md
+     - Clustering Election Discourse on Social Media: tutorials/elections.md
+     - Dynamic Time Series Analysis of Financial News: tutorials/financial.md
   - Models:
     - Model Overview: model_overview.md
     - Semantic Signal Separation (S³): s3.md
