+++ docs/tutorials/religious.md
@@ -1,0 +1,166 @@
+# Analyzing Discourse on Religion and Morality
+
+Topic models are an effective tool for discourse analysis, and widely applied in computational humanities research.
+One potential research question could be to investigate how groups online discuss the connection between morality and religion.
+
+For this task, we are going to utilize [KeyNMF](../KeyNMF.md), which is a powerful topic model based on keyword/keyphrase extraction using contextual representations.
+
+## Installation
+
+```bash
+pip install turftopic[topic-wizard]
+```
+
+## Data Preparation
+
+We are going to use a subset of the 20 Newsgroups dataset, which includes three newsgroups oriented at discussing religion and atheism.
+Luckily this dataset is directly available from scikit-learn's repositories.
+
+```python
+import numpy as np
+from sklearn.datasets import fetch_20newsgroups
+
+ds = fetch_20newsgroups(
+    subset="all",
+    remove=("headers", "footers", "quotes"),
+    categories=[
+        "alt.atheism",
+        "talk.religion.misc",
+        "soc.religion.christian",
+    ],
+)
+corpus = ds.data
+group_labels = np.array(ds.target_names)[ds.target]
+```
+
+Turftopic uses contextual embeddings of documents in order to understand what is in the corpus.
+We will use the small and fast `all-MiniLM-L6-v2` sentence transformer to produce embeddings of our data.
+
+!!! tip 
+    Sometimes you might want to reuse embeddings in different topic models, or save them to disk.
+    It is thus recommended, but not necessary to precompute them before running a topic model.
+
+```python
+from sentence_transformers import SentenceTransformer
+
+encoder = SentenceTransformer("all-MiniLM-L6-v2")
+embeddings = encoder.encode(corpus, show_progress_bar=True)
+```
+
+## Model Training
+
+Since we want to investigate the discourse from the perspective of morality, we would like to include this information in our model.
+KeyNMF is capable of investigating a corpus from a certain angle based on a **seed phrase**, this will allow us to focus the topics in the model on morality.
+
+```python
+from turftopic import KeyNMF
+
+model = KeyNMF(
+    n_components=15,
+    random_state=42,
+    encoder=encoder,
+    seed_phrase="Religion and Morality"
+)
+topic_data = model.prepare_topic_data(corpus, embeddings=embeddings)
+```
+
+## Model Interpretation
+
+Let us first print the top 10 words for each topic discovered by the model.
+
+```python
+topic_data.print_topics()
+```
+
+| Topic ID | Highest Ranking |
+| - | - |
+| 0 | atheist, atheists, theists, religious, abortion, agnostic, theist, just, mythology, argument |
+| 1 | immoral, morals, morality, moral, morally, behavior, society, religious, meat, societies |
+| 2 | bible, scripture, biblical, scriptures, faith, gospel, testament, commandments, revelation, interpretation |
+| 3 | religion, religious, religions, beliefs, cult, science, spiritual, don, scientific, secular |
+| 4 | god, believe, satan, existence, gods, evil, exist, just, genocide, creator |
+| 5 | homosexuality, homosexual, homosexuals, heterosexual, sexual, gay, fornication, immoral, sex, sodom |
+| 6 | morality, objective, subjective, absolute, natural, relativism, societal, objectively, science, correct |
+| 7 | belief, faith, beliefs, believe, believing, evidence, philosophy, religions, believed, believes |
+| 8 | atheism, theism, religious, alt, belief, stalin, agnosticism, theists, newsgroup, read |
+| 9 | christians, religions, jesus, fundamentalist, beliefs, commandments, hell, worship, jewish, christian |
+| 10 | christianity, christian, christ, gospel, evangelical, cult, book, life, faith, buddhist |
[…]
