### 2025-05-21 #8183 @Jemoka
@x-tabdeveloping

Looks like your dependencies are listed here using Poetry 1.x standards, which won't install for folks running any other package manager who wants to build from source.

For the installation instructions in the JOSS checklist, could you migrate either to the new [PEP 517](https://peps.python.org/pep-0517/) standard across Python, which [Poetry 2.x recommends](https://github.com/python-poetry/poetry/blob/84eeadc21f92a04d46ea769e3e39d7c902e44136/docs/dependency-specification.md), or otherwise generating a requirements.txt?

https://github.com/x-tabdeveloping/turftopic/issues/94
