### 2025-05-21 #8183 @Jemoka
currently, the example code in the README reads:

```python
from sklearn.datasets import fetch_20newsgroups
newsgroups = fetch_20newsgroups(
    subset="all",
    remove=("headers", "footers", "quotes"),
)
corpus = newsgroups.data

#....

from turftopic import KeyNMF
model = KeyNMF(20).fit(corpus)
```

which doesn't give us a lot of information about how `corpus` is supposed to look as input. Clarifying this would help folks get started quickly, for JOSS' "Example usage" requirements.

https://github.com/x-tabdeveloping/turftopic/issues/95
