### 2025-05-21 #8183 @Jemoka
Echoing @PetrKorab's point, I'd love to see more extended in the Use Cases and Target Audience sections. In particular, as noted in the statement of need, there's already a few competing packages that does only a part or some models in topic modeling. What would this package do differently to help community adoption (for instance, for which scenarios that would need multiple baselines or benchmarks? or perhaps how it could be more extensible than others? etc.)
