### 2025-05-21 x-tabdeveloping/turftopic#93 @PetrKorab
I am sorry for not spotting this section. Yes, it is well-explained. I trained a KeyNMF model with a custom bigram vectorizer, and it ran well.

Just a small point: It seems that the markdown in these pages does not display the notion well:

https://x-tabdeveloping.github.io/turftopic
https://x-tabdeveloping.github.io/turftopic/model_overview

Eg. | | | | | - | - | - | | :house: [Build and Train Topic Models](model_definition_and_training.md) | :art: [Explore, Interpret and Visualize your Models](model_interpretation.md) | :wrench: [Modify and Fine-tune Topic Models](finetuning.md) | | :pushpin: [Choose the Right Model for your Use-Case](model_overview.md) | :chart_with_upwards_trend: [Explore Topics Changing over Time](dynamic.md) | :newspaper: [Use Phrases or Lemmas for Topic Models](vectorizers.md) | | :ocean: [Extract Topics from a Stream of Documents](online.md) | :evergreen_tree: [Find Hierarchical Order in Topics](hierarchical.md) | :whale: [Name Topics with Large Language Models](namers.md) |

Thanks for making it better readable!
