### 2025-05-21 #8183 @x-tabdeveloping
Hi @Jemoka ! Thanks for the feedback. I already started working on @PetrKorab 's suggestion regarding expanding use cases, and I am in the middle of writing a couple of tutorials in the documentation for various use cases. I will also adjust the paper, and keep you in the loop.

Thanks for catching the problem with the dependency specification, I somehow missed this.

Regarding the model comparison scenario that you mentioned: We have so far not added metrics to the library, as, as far as I can tell, there isn't general consensus in the topic modelling literature on whether metrics are a good measure of topic quality, and if so, which ones. The topic models in our library also conceptualize topics in a vast array of different ways. For our publications we have so far used the [topic-benchmark](https://github.com/x-tabdeveloping/topic-benchmark) library, where I have implemented a number of metrics.
Do you think Turftopic itself would benefit from adding metrics of topic quality?
