### 2025-05-21 #8183 @Jemoka
> The topic models in our library also conceptualize topics in a vast array of different ways. For our publications we have so far used the [topic-benchmark](https://github.com/x-tabdeveloping/topic-benchmark) library, where I have implemented a number of metrics.
> Do you think Turftopic itself would benefit from adding metrics of topic quality?

Yes! Only because of the fact that you are releasing, per article, the first known implementations of a few SOTA models. I do think having some benchmarks and allowing a pipeline for future contributions to benchmarked would be great. In particular, I think this would make yours the go-to implementation of some of these approaches, otherwise its just another topic modeling library and folks can use BertTopic or others instead :)
