### 2025-05-26 x-tabdeveloping/turftopic#96 @x-tabdeveloping
Add metrics and tools for model comparison
As per suggested by @Jemoka:

Adding utilities for comparing topic models would make the library stand out from the rest of the crowd, and would make a more convincing point for adoption by most people.

I think the following would make quite a bit of sense:
1. Embedding-based coherence on:
   1. Words
   2. Documents
   3. Images
2. Diversity metrics:
   1. Embedding-based
   2. Lexical

It would also make a lot of sense to add the option for significance tests, since most of these metrics are aggregated.
