### 2025-05-26 #8183 @x-tabdeveloping
Sorry for the radio silence, I got ill last week, and needed to take a couple of days off.
I have made an [issue on model comparison](https://github.com/x-tabdeveloping/turftopic/issues/96) as @Jemoka requested. I will get to it as soon as I am done with the use cases.

In terms of the use cases section, I have already written two tutorials on [this branch](https://github.com/x-tabdeveloping/turftopic/tree/joss_use_cases) (Building a taxonomy of machine learning papers; Discourse analysis on morality and religion), and I'm currently working on a third. I will tag you in the upcoming pull request once I'm done with them.
