### 2025-05-26 79a70b9
Added political compass tutorial
