+++ docs/tutorials/ideologies.md
@@ -1,0 +1,172 @@
+# Discovering a Data-driven Political Compass
+
+The Political Compass is a dimensional theory of political ideologies and views.
+This model posits that political ideology is distributed along a Left-Right and Libertarian-Authoritarian axis.
+
+<figure>
+  <img src="https://upload.wikimedia.org/wikipedia/commons/thumb/5/5c/Political_Compass_purple_LibRight.svg/1920px-Political_Compass_purple_LibRight.svg.png" width="400px" style="margin-left: auto;margin-right: auto;">
+  <figcaption>The Political Compass <br>(figure from Political Compass website)</figcaption>
+</figure>
+
+While this model enjoys wide public recognition, one potential issue with it is that it is a **top-down** model,
+meaning that these dimensions were not discovered from some underlying data, but is based on experts' intuitions.
+Dimensional analysis of views is also typically conducted using surveys.
+
+In this tutorial we are going to look into how one could discover a **bottom-up**, data-driven Political Compass using the power of topic modelling.
+
+## Installation
+
+We will install Turftopic with Plotly to be able to plot our results, and the `datasets` library, for fetching data from HF Hub.
+
+```python
+pip install datasets plotly pandas turftopic
+```
+
+## Data Preparation
+
+For this demonstration, I will be using a synthetic dataset, in which a large language model was tasked with expressing political opinions in free-form text.
+
+```python
+from datasets import load_dataset
+
+ds = load_dataset("JyotiNayak/political_ideologies", split="train")
+texts = ds["statement"]
+```
+
+We will be using the `paraphrase-MiniLM-L12-v2` for embedding our dataset and pre-computing embeddings.
+
+```python
+from sentence_transformers import SentenceTransformer
+
+encoder = SentenceTransformer("paraphrase-MiniLM-L12-v2")
+embeddings = encoder.encode(texts, show_progress_bar=True)
+```
+
+## Model Training
+
+We will use the $S^3$ topic model for our investigations, as it conceptualized topics as independent axes in semantic space,
+meaning it is built for establishing dimensional theories similar to the Political Compass.
+For more details, read the documentation page on [Semantic Signal Separation](../s3.md).
+
+Instead of a 2-dimensional model, similar to the Political Compass, we will opt to discover 3 dimensions.
+
+!!! note
+    You can easily expand this to more dimensions, the only reason we're not doing it here is because it would take more time to interpret them, and the tutorial is more accessible this way.
+
+```python
+from turftopic import SemanticSignalSeparation
+
+model = SemanticSignalSeparation(3, encoder=encoder, random_state=42)
+doc_topic_matrix = model.fit_transform(texts, embeddings=embeddings)
+```
+
+## Model Interpretation
+
+First, let us examine the highest and lowest ranking terms on each axis in order to gain an intuition for what the dimensions could be about.
+
+```python
+model.print_topics(top_k=10)
+```
+
+| Topic ID | Highest Ranking | Lowest Ranking |
+| - | - | - |
+| 0 | religion, religious, faith, church, religions, faiths, doctrines, freedom, freedoms, beliefs | households, labor, household, poverty, socioeconomic, hardworking, income, wage, pay, welfare |
+| 1 | investments, investment, spending, fiscal, invest, funding, policy, pollution, economic, budget | racism, racial, ethnicity, diverse, discrimination, distinct, genders , ethnic, families, adoption |
+| 2 | warming, carbon, environment, environmental, planet, change, solar, greenhouse, fossil, biodiversity | wealth, taxation, prosperity, wealthiest, tax, profit, entrepreneurship, taxes, fiscal, government |
+
[…]
