+++ docs/tutorials/reviews.md
@@ -1,0 +1,129 @@
+# Customer Dissatisfaction Analysis
+
+When providing some service for customers as a company, it is valuable to have an overview of what customers are dissatisfied with.
+A good place to start would be to analyze reviews of your product that have low scores.
+
+In this tutorial we will use Turftopic to find topics in customer reviews of the Uber app on Google Play.
+The dataset is [openly available on Kaggle](https://www.kaggle.com/datasets/kanchana1990/uber-customer-reviews-dataset-2024?resource=download).
+
+## Installation
+
+We will be using SpaCy for phrase extraction and Plotly for producing visualizations and pandas for data wrangling.
+We will also have to install a SpaCy pipeline.
+
+```python
+pip install turftopic[spacy] plotly pandas
+python -m spacy download en_core_web_sm
+```
+
+## Data Preparation
+
+We will load the dataset, and extract the reviews with the lowest scores, to see what the most disappointed users take issue with.
+
+```python
+import pandas as pd
+
+df = pd.read_csv("uber_reviews_without_reviewid.csv")
+corpus = list(df["content"][df["score"] == 1])
+```
+
+Since the dataset contains reviews in Spanish, we will be using a multilingual encoder model:
+
+```python
+from sentence_transformers import SentenceTransformer
+
+encoder = SentenceTransformer("paraphrase-multilingual-MiniLM-L12-v2")
+embeddings = encoder.encode(corpus, show_progress_bar=True)
+```
+
+## Model Training
+
+We will be using the [KeyNMF](../KeyNMF.md) model for topic discovery, as it is capable of hierarchically expanding topics if we need it.
+
+!!! tip
+    Using noun phrases, as it is done in this tutorial can greatly increase the interpretability, clarity and precision of your topic descriptions, but can make model training substantially slower, as sentences have to be parsed.
+    See the [Vectorizers](../vectorizers.md) page for more detail.
+
+```python
+from turftopic import KeyNMF
+from turftopic.vectorizers.spacy import NounPhraseCountVectorizer
+
+model = KeyNMF(
+    # We train first with 10 topics, and can then expand the hierarchy if needed
+    n_components=10,
+    encoder=encoder,
+    vectorizer=NounPhraseCountVectorizer("en_core_web_sm"),
+    random_state=42,
+)
+doc_topic_matrix = model.fit_transform(corpus, embeddings=embeddings)
+```
+
+## Model Interpretation
+
+Let us first examine the highest ranking phrases on each topic.
+
+```python
+model.print_topics()
+```
+
+| Topic ID | Highest Ranking |
+| - | - |
+| 0 | driver, pickup, cancellation fee, car, money, fee, location, extra money, road, auto driver |
+| 1 | app, apps, worst application, option, taxi, phone, support, ads, time, customer service |
+| 2 | drivers, rides, company, time, trips, customer, riders, customers, car, apps |
+| 3 | ride, time, rider, charge, destination, worst service, rides, emergency no rider, 2 rides, uber app |
+| 4 | worst app, booking, customer support, waiting time, destination, rides, scam, worst customer support, senseless irresponsible app, traffic useless app |
+| 5 | cab, cab driver, time, emergency, booking, auto, bad service, worst cab bookinv app, rapido, amount |
+| 6 | account, money, customer support, refund, card, bank, permission, email, credit card, customers |
+| 7 | price, prices, pricing, fees, traffic, predatory pricing, destination, higher price, discount, amount |
+| 8 | trip, cash, amount, car, customer support, service, trips, bad experience, destination, cancelation fee |
+| 9 | uber, service, payment, taxi, rides, customer service, worst service, uber app, customers, refund |
+
[…]
