+++ docs/tutorials/overview.md
@@ -1,0 +1,24 @@
+## Case Studies
+
+Topic models can be used in various real-world scenarios in both academia and industry.
+We provide a number of concrete examples of using Turftopic to gain real insights into the nature of text data.
+
+<div class="grid" markdown>
+
+<img src="../images/preview_cluster.png" height="500px"/>
+:fontawesome-solid-circle-nodes: [**Cluster Analysis** of the landscape of machine learning research.](./arxiv_ml.md)
+{ .card }
+
+<img src="../images/preview_word_map.png" height="500px"/>
+:fontawesome-solid-church: [**Discourse Analysis** of internet forums on morality and religion.](./religious.md)
+{ .card }
+
+<img src="../images/preview_pol_comp.png" height="500px"/>
+:fontawesome-solid-compass: [**Dimensional Analysis** of political ideologies and issues.](./ideologies.md)
+{ .card }
+
+<img src="../images/preview_customer.png" height="500px"/>
+:fontawesome-solid-car-side: [**Dissatisfaction Analysis** of customer reviews.](./reviews.md)
+{ .card }
+
+</div>
