### 2025-05-26 x-tabdeveloping/turftopic#93 @x-tabdeveloping
I will be closing this issue, since only the markdown table problem is still present.
