+++ paper.bib
@@ -178,3 +178,87 @@
   version = {0.5.0},
   year = {2023}
 }
+
+@article{discourse_analysis,
+  author = {Thomas Jacobs and Robin Tschötschel and},
+  title = {Topic models meet discourse analysis: a quantitative tool for a qualitative approach},
+  journal = {International Journal of Social Research Methodology},
+  volume = {22},
+  number = {5},
+  pages = {469--485},
+  year = {2019},
+  publisher = {Routledge},
+  doi = {10.1080/13645579.2019.1576317},
+}
+
+@book{macroanalysis,
+  ISBN = {9780252037528},
+  URL = {http://www.jstor.org/stable/10.5406/j.ctt2jcc3m},
+  abstract = {In this volume, Matthew L. Jockers introduces readers to large-scale literary computing and the revolutionary potential of macroanalysis--a new approach to the study of the literary record designed for probing the digital-textual world as it exists today, in digital form and in large quantities. Using computational analysis to retrieve key words, phrases, and linguistic patterns across thousands of texts in digital libraries, researchers can draw conclusions based on quantifiable evidence regarding how literary trends are employed over time, across periods, within regions, or within demographic groups, as well as how cultural, historical, and societal linkages may bind individual authors, texts, and genres into an aggregate literary culture. Moving beyond the limitations of literary interpretation based on the close-reading of individual works, Jockers describes how this new method of studying large collections of digital material can help us to better understand and contextualize the individual works within those collections.},
+  author = {MATTHEW L. JOCKERS},
+  publisher = {University of Illinois Press},
+  title = {Macroanalysis: Digital Methods and Literary History},
+  urldate = {2025-05-27},
+  year = {2013}
+}
+
+@article{hotel_sector, title={Analysing online customer experience in hotel sector using dynamic topic modelling and net promoter score}, volume={14}, url={https://www.emerald.com/insight/content/doi/10.1108/jhtt-04-2021-0116/full/html}, DOI={10.1108/jhtt-04-2021-0116}, number={2}, journal={Journal of Hospitality and Tourism Technology}, author={Nguyen, Van-Ho and Ho, Thanh}, year={2023}, month=feb, pages={258–277} }
+
+@article{social_media_mining, title={Mining social media data via supervised topic model: Can social media posts inform customer satisfaction?}, url={https://onlinelibrary.wiley.com/doi/full/10.1111/deci.12660}, DOI={10.1111/deci.12660}, journal={Decision Sciences}, author={Huang, Yinghui and Li, Mei and Tsung, Fugee and Chang, Xiangyu}, year={2025}, month=jan }
+
+@InProceedings{content_recommendation,
+  author="Bergamaschi, Sonia
+  and Po, Laura",
+  editor="Monfort, Val{\'e}rie
+  and Krempels, Karl-Heinz",
+  title="Comparing LDA and LSA Topic Models for Content-Based Movie Recommendation Systems",
+  booktitle="Web Information Systems and Technologies",
+  year="2015",
+  publisher="Springer International Publishing",
+  address="Cham",
+  pages="247--263",
[…]
