### 2025-05-27 x-tabdeveloping/turftopic#96 @KennethEnevoldsen
@x-tabdeveloping, can you outline what you think is the desirable user interface here? Then we can just discuss from there?

(I could imagine that people might want to compare models visually down the line as well, but I don't think this is now)
