### 2025-05-28 a6bd5cc
Update docs/tutorials/arxiv_ml.md
Co-authored-by: Kenneth Enevoldsen <kennethcenevoldsen@gmail.com>
