### 2025-05-28 000f89d
Adjusted tutorials according to feedback
