+++ .github/workflows/documentation.yml
@@ -23,7 +23,7 @@ jobs:
       - name: Dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install "turftopic[pyro-ppl]" "griffe" "mkdocstrings[python]" "mkdocs" "mkdocs-material"
+          pip install ".[dev]"
 
       - name: Build and Deploy
         if: github.event_name == 'push'
