+++ pyproject.toml
@@ -18,7 +18,6 @@ license = { text = "MIT" }
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
-  "python>=3.10",
   "numpy>=1.23.0",
   "scikit-learn>=1.3.0,<2.0.0",
   "sentence-transformers>=2.2.0",
@@ -37,7 +36,7 @@ openai = ["openai>=1.40.0,<2.0.0"]
 opentsne = ["openTSNE>=1.0.0,<2.0.0"]
 datamapplot=["datamapplot>=0.4.2, <1.0.0"]
 jieba = ["jieba>=0.40.0,<1.0.0"]
-spacy = ["spacy>=3.6.0,<1.0.0"]
+spacy = ["spacy>=3.6.0,<4.0.0"]
 snowball = ["snowballstemmer>=2.0.0,<3.0.0"]
 topic-wizard = ["topic-wizard>1.0.0,<2.0.0"]
 umap-learn = ["umap-learn>=0.5.5,<1.0.0"]
@@ -55,7 +54,6 @@ dev = [
   "openTSNE>=1.0.0,<2.0.0",
   "datamapplot>=0.4.2, <1.0.0",
   "jieba>=0.40.0,<1.0.0",
-  "spacy>=3.6.0,<1.0.0",
   "snowballstemmer>=2.0.0,<3.0.0",
   "topic-wizard>1.0.0,<2.0.0",
   "umap-learn>=0.5.5,<1.0.0",
