+++ pyproject.toml
@@ -41,6 +41,7 @@ snowball = ["snowballstemmer>=2.0.0,<3.0.0"]
 topic-wizard = ["topic-wizard>1.0.0,<2.0.0"]
 umap-learn = ["umap-learn>=0.5.5,<1.0.0"]
 docs = [
+  "griffe==0.40.0",
   "mkdocs==1.5.3",
   "mkdocs-autorefs==0.5.0",
   "mkdocs-material==9.5.6",
@@ -63,6 +64,7 @@ dev = [
   "mkdocs-material-extensions==1.3.1",
   "mkdocstrings==0.22.0",
   "mkdocstrings-python==1.8.0",
+  "griffe==0.40.0",
 ]
 
 [build-system]
