### 2025-05-29 #8183 @PetrKorab
The docs are nicely updated with tutorials. The Use Cases and Target Audience sections of the paper are extended. Closed https://github.com/x-tabdeveloping/turftopic/issues/92.
