### 2025-05-29 x-tabdeveloping/turftopic#100 @PetrKorab
Add automated tests - JOSS
For JOSS, the authors are expected to provide automated tests to demonstrate the software's functionality. These can be GitHub actions or other CI (Travis CI, ..). The best might be to use the examples in the Tutorials section for testing.

See more:  https://joss.readthedocs.io/en/latest/submitting.html.

Please, make them accessible to the reviewers on the repo. Thanks!
