### 2025-05-29 #8183 @PetrKorab
@x-tabdeveloping For JOSS, the authors are expected to provide automated tests to demonstrate the software's functionality. These can be GitHub actions or other CI (Travis CI, ..). The best might be to use the examples in the Tutorials section for testing.

See more: https://joss.readthedocs.io/en/latest/submitting.html.

Please, make them accessible to the reviewers on the repo. Thanks!

https://github.com/x-tabdeveloping/turftopic/issues/100
