### 2025-05-29 x-tabdeveloping/turftopic#95 @Jemoka
could you extend the README to clarify the input type (i.e. what should `corpus` look like)? this allows example usage to be surfaced easily.
