### 2025-05-29 x-tabdeveloping/turftopic#95 @x-tabdeveloping
I might be missing something but I thought I did that. The readme now has type annotation on the corpus variable
