### 2025-06-03 faba545
Added fighting words to clustering models and made linkage more numerically stable
