### 2025-06-03 ebe1b3a
Added semi-supervised modelling for clustering models, and made BERTopic and Top2Vec more flexible
