+++ turftopic/models/cluster.py
@@ -100,9 +100,19 @@ def build_tsne(*args, **kwargs):
     try:
         from openTSNE import TSNE
 
-        model = TSNE(*args, **kwargs)
-        model.fit_transform = model.fit
-        return model
+        class OpenTSNEWrapper(TSNE):
+            def __init__(self, *args, **kwargs):
+                super().__init__(*args, **kwargs)
+
+            def fit_transform(self, X: np.ndarray, y=None):
+                return super().fit(X)
+
+            def fit(self, X: np.ndarray, y=None):
+                self.fit_transform(X, y)
+                return self
+
+        return OpenTSNEWrapper(*args, **kwargs)
+
     except ModuleNotFoundError:
         from sklearn.manifold import TSNE
 
@@ -434,7 +444,8 @@ class ClusteringTopicModel(
         raw_documents: iterable of str
             Documents to fit the model on.
         y: None
-            Ignored, exists for sklearn compatibility.
+            Originally ignored, in case of a dimensionality reduction that can utilize labels,
+            you can pass labels to the model to inform the clustering process.
         embeddings: ndarray of shape (n_documents, n_dimensions), optional
             Precomputed document encodings.
 
@@ -454,8 +465,12 @@ class ClusteringTopicModel(
             self.doc_term_matrix = self.vectorizer.fit_transform(raw_documents)
             console.log("Term extraction done.")
             status.update("Reducing Dimensionality")
+            # If y is specified, we pass it to the dimensionality
+            # reduction method as supervisory signal
+            if y is not None:
+                _, y = np.unique(y, return_inverse=True)
             self.reduced_embeddings = (
-                self.dimensionality_reduction.fit_transform(embeddings)
+                self.dimensionality_reduction.fit_transform(embeddings, y=y)
             )
             console.log("Dimensionality reduction done.")
             status.update("Clustering documents")
@@ -526,6 +541,7 @@ class ClusteringTopicModel(
         doc_topic_matrix = self.fit_transform(
             raw_documents,
             embeddings=self.multimodal_embeddings["document_embeddings"],
+            y=y,
         )
         self.image_topic_matrix = self.transform(
             raw_documents,
@@ -720,37 +736,21 @@ class ClusteringTopicModel(
 
 class BERTopic(ClusteringTopicModel):
     """Convenience function to construct a BERTopic model in Turftopic.
+    The model is essentially just a ClusteringTopicModel
+    with BERTopic's defaults (UMAP -> HDBSCAN -> C-TF-IDF).
+
+    ```bash
+    pip install turftopic[umap-learn]
+    ```
 
     ```python
     from turftopic import BERTopic
-    from sklearn.cluster import HDBSCAN
-    import umap
 
     corpus: list[str] = ["some text", "more text", ...]
 
     model = BERTopic().fit(corpus)
     model.print_topics()
     ```
-
-    Parameters
-    ----------
-    encoder: str or SentenceTransformer
-        Model to encode documents/terms, all-MiniLM-L6-v2 is the default.
-    vectorizer: CountVectorizer, default None
-        Vectorizer used for term extraction.
-        Can be used to prune or filter the vocabulary.
-    dimensionality_reduction: TransformerMixin, default None
-        Dimensionality reduction step to run before clustering.
-        Defaults to UMAP(5, metric="cosine")
-    clustering: ClusterMixin, default None
-        Clustering method to use for finding topics.
-        Defaults to HDBSCAN.
-    n_reduce_to: int, default None
-        Number of topics to reduce topics to.
-        The specified reduction method will be used to merge them.
-        By default, topics are not merged.
-    random_state: int, default None
-        Random state to use so that results are exactly reproducible.
     """
 
     def __init__(
[…]
