### 2025-06-03 6b7fa4a
Added proper label factorization for semi-supervised models
