### 2025-06-13 #8183 @marccanby
First of all, great work! I think this is an interesting and useful package. Your documentation is especially good.

**Draft**
I think you should say what topic modeling is at the very top (ability to automatically extract topics from a set of documents), so that its usefulness may be clear to a reader who is not familiar with the topic; this is especially important for this kind of work because I feel that it is designed to be easily accessible to people who aren’t as computationally oriented.

In the draft you say that “we included an extensive suite of pretty-printing and visualization utilities”, but the README makes it clear that “Turftopic does not come with built-in visualization utilities, [topicwizard](https://github.com/x-tabdeveloping/topicwizard), an interactive topic model visualization library, is compatible with all models from Turftopic.” I feel that the sentence in your draft is misleading. Can you instead emphasize the compatibility of your library with existing visualization libraries (like topicwizard), rather than make it sound like you implemented visualizations in your library?

I think the draft would benefit if it included a brief example to make things more concrete. This could help show how the library is useful, in particular with regard to how it is distinguished from similar tools. Perhaps show a very simple workflow of learning a topic model and visualizing the topics, emphasizing how easy it is to switch in and out the topic models?

Why do some of your citations say “et al.” after already providing a list of authors? For example “Bianchi, Terragni, Hovy, Nozza, et al., 2021” should just be “Bianchi et al., 2021”.

**README**
For zsh users (e.g. on Mac), the command `pip install turftopic[pyro-ppl]` does not work because of the square brackets - but `pip install "turftopic[pyro-ppl]”` works. Can you change the README to include the quotes?

I had issues on the line `newsgroups = fetch_20newsgroups(subset="all", remove=("headers", "footers", "quotes”))` because I didn’t have the required certificates installed. I think it could be better if the README said something like “(for Mac users, if you encounter SSL certificate errors, you’ll need to install the required certificates first)” - perhaps also with the command to do so.

This line is an issue: `model.print_representative_documents(0, corpus, document_topic_matrix)` because `document_topic_matrix` was not previously defined. I believe it should be `model.print_representative_documents(0, corpus, model.document_topic_matrix)`. (Although, if this is a method of `model`, then I’m not sure why you have to pass an argument to it that it already has in its instance variables.)

When I run `namer = OpenAITopicNamer("gpt-4o-mini”)`, it says I need to install `openai`. Can you clarify that in the README? You should also say that running that line requires an API key. A similar installation issue occurs with spacy in the next section.

**Some very minor comments in the draft:**
- I think “which users are likely already familiar with” should instead be “which many users may be familiar with”, especially if this is targeted at a broad audience (that way it doesn’t seem like you make an assumption about the readers’ technical skills)
- The sentence that starts “The library also includes three topic models, which...” is confusing, because your library has more than 3 models - it should instead say “The library also includes three topic models that, to our knowledge, only have implementations in Turftopic:...” (Basically, make it a restrictive relative clause)
