### 2025-06-16 aded872
Adjusted readme according to feedback
