### 2025-06-16 bd0eb49
Added more details on visualization in readme
