### 2025-06-17 #8183 @abhishektiwari
@x-tabdeveloping yes example should be part of repository and/or software documentation. Also, we would like to keep JOSS paper less than 1000 words. You can always link examples in paper if needed.
