### 2025-06-18 #8183 @x-tabdeveloping
Hi @abhishektiwari . Of course:
1. I have double checked the authors and affiliations
2. Made a release (0.17.2), DOI: [10.5281/zenodo.15688293](https://doi.org/10.5281/zenodo.15688292)
3. I have made sure the authors, OrcIDs and affiliations are consistent with the paper
4. I have checked that the license is the same (MIT)
