### 2025-06-19 6f78628
Rephrased documentation and readded formulae
