### 2025-06-19 d38bcfe
Rephrased warning message
