### 2025-06-20 #8183 @abhishektiwari
@x-tabdeveloping Zenodo archive author metadata is slightly different from paper.

Here is citation from the Zenodo archive,

Márton Kardos, Enevoldsen, K. C., Kostkan, J., Rocca, R., & Kristensen-McLachlan, R. D. (2025). x-tabdeveloping/turftopic: v.0.17.2 (v0.17.2). Zenodo. https://doi.org/10.5281/zenodo.15688293

First it should be Kardos, Márton. Secondly author order should be in same order as paper.

Can you update archive and share DOI?
