### 2025-06-24 #8183 @abhishektiwari
@x-tabdeveloping I still see differences between author metadata in Paper vs. Zenodo archive. For example ORCID for [Rocca, Roberta](https://zenodo.org/search?q=metadata.creators.person_or_org.name:%22Rocca,+Roberta%22) is different. Please review.

In Zenodo https://orcid.org/0000-0001-9017-8088
In Paper https://orcid.org/0000-0002-0680-7097
