### 2025-07-01 #8183 @abhishektiwari
@x-tabdeveloping Can you please update title to [Zenodo](https://zenodo.org/records/15688293) to match with paper?

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
