### 2025-07-01 d6bb74f
Update paper.md for a minor grammatical improvements
Update paper.md for a minor grammatical improvements
