+++ paper.bib
@@ -10,7 +10,7 @@
 }
 
 @inproceedings{keynmf,
-    title = "Context is Key(NMF):: Modelling Topical Information Dynamics in Chinese Diaspora Media",
+    title = "Context is Key(NMF):: Modelling Topical Information Dynamics in {C}hinese Diaspora Media",
     abstract = "Does the People{\textquoteright}s Republic of China (PRC) interfere with European elections through ethnic Chinese diaspora media? This question forms the basis of an ongoing research project exploring how PRC narratives about European elections are represented in Chinese diaspora media, and thus the objectives of PRC news media manipulation. In order to study diaspora media ef{\"I}ciently and at scale, it is necessary to use techniques derived from quantitative text analysis, such as topic modelling. In this paper, we present a pipeline for studying information dynamics in Chinese media. Firstly, we present KeyNMF, a new approach to static and dynamic topic modelling using transformer-based contextual embedding models. We provide benchmark evaluations to demonstrate that our approach is competitive on a number of Chinese datasets and metrics. Secondly, we integrate KeyNMF with existing methods for describing information dynamics in complex systems. We apply this pipeline to data from five news sites, focusing on the period of time leading up to the 2024 European parliamentary elections. Our methods and results demonstrate the effectiveness of KeyNMF for studying information dynamics in Chinese media and lay groundwork for further work addressing the broader research questions.",
     keywords = "Chinese, contextual topic models, information dynamics, keywords, novelty",
     author = "Kristensen-McLachlan, {Ross Deans} and Hicke, {Rebecca Marie Matouschek} and M{\'a}rton Kardos and Mette Thun{\o}",
