### 2025-07-02 #8183 @x-tabdeveloping
Hi @crvernon ! Thanks for pointing that out! I have now updated the bib file and wrapped the C in curly brackets.
