### 2025-07-03 #8183 @crvernon
🥳 Congratulations on your new publication @x-tabdeveloping! Many thanks to @abhishektiwari for editing and @PetrKorab, @Jemoka and @marccanby for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
